repair: Move repair_writer to separate header.
